Remove percentage limit for new assignments.
As the limit value for the percentage evaluation is not yet used, it is
probably better to spare the users from setting it (in vain).  Therefore it is
always implicitly defaulting to 0%.  Add a label for the selection as well,
while I am here.

Note that, despite it becomes a static selection this way, the Fay binding is
left there for now.  One could possibly remove this at a later occasion.

Also note that now it is possible to change the evalution type of an
assignment even there are submissions for it.  This option shall be disabled
after the first submission was submitted in order to maintain consistency.
However, apparently it does not cause any harm during the operation.