Add a utility script for simulating the testing backend.
This can be used by the course administrators to develop their own tester
scripts.  It is recommended to setup a dedicated system for this purpose,
install this script there, and hand out access for the CAs so they could
login to there and interactively experiment with their scripts.  It shall be
compatible with the backend.