Disallow communicating non-ASCII or non-UTF-8 output to the frontend.
The frontend expects the comments generated by the backend are all either
plain ASCII or Unicode (UTF-8) text.  Otherwise an exception is raised in
the feedback manager thread every time it sees the offending comment and
thus the processing stucks.

This is now worked around by replacing the actual comment with a warning
that informs the user about his or her wrongdoing.