Add support for stacking dictionaries.
This is to provide a way to apply only minor modifications to the
dictionaries that have been already created.  That way users can easily
customize the user interface depending on the needs of the actual use.

There is some error handling also implemented, so the server can emit
messages (and refuse to start) when the detected dictionaries or patches
do something wrong.