Switch to TemplateHaskell for handling dictionary entries.
This is a preparation for the upcoming commit.  Its purpose is to automatize
assignment of (unique) identifiers for each of the translation entries with
the use of Template Haskell.  It also helps with assigning labels of the same
name, so they could be tracked back much easier when something does not work
properly.