Add a Markdown handler.
This handler makes it possible to place pure Markdown files under the
"markdown" directory and serve them as HTML pages under the "/markdown/"
link.  However, note that the pages cannot be standalone, but they will be
displayed in frames with the global Bootstrap header.  It is mostly
recommended for public information pages or documentation whose source could
also be maintained in the git repository, in easily editable Markdown format.