Make it possible to patch the default (built-in) language messages.
It turned out that the current implementation does not let the user to load
dictionaries or dictionary patches for the language (which is currently "en")
that is implictly contained in the sources.  This limitation is now removed,
and a `defaultDictionary` constant is exported from `Bead.View.Dictionary`
that could be referenced as the parent for the respective patches.