Expand size of text fields in SQL to avoid unintentional cut-offs.
Unfortunately, the default "text" fields in MySQL can only be of size 2^16
(that is, 64 KB) at maximum, which appears to be quite small for certain
fields stored in the database.  Thus they are now changed to "longtext"
so they could range up to 2^32, which should be enough for a while.  Note
that the Unicode character setting has to be also added because
`persistent-mysql` does not know anything about longtexts (at the moment).