Work around `file(1)` checks for text files starting with `"Seg..."`.
The `file(1)` utility (on FreeBSD, at least) is prone to identify the contents
of a text file badly when it starts with the word `"Segmentation"` (as in the
typical error message, `"Segmentation fault (core dumped)"`).  In that case,
it is considered `"StuffIt Deluxe Segment (data)"` (possibly due to the
initials `"Seg"`) instead of `"ASCII text"`.