Use utf8mb4 for every text field in the SQL database.
It turns out that the default Unicode encoding in MySQL, utf8 does not
properly support certain characters (such as the ones that are encoded in 4
bytes).  So when a string used at insertion or update contains them, MySQL
simply cuts off the contents, which basically leads to unintented data
loss.

For more details, see https://mathiasbynens.be/notes/mysql-utf8mb4