Fix saving new score to avoid duplication.
It was possible to create two scores for a user and assessment,
thus voilating the assumption that each user has at most one score
for each assessment. On the user's side, the visible effect was
that the assessment disappeared.

Now it is verified that there is no score for a user before saving
a new score.