Make it possible to see feedbacks for submissions in ballot box mode.
Previously, it was not allowed to access the submission details page
when ballot box mode has been actived.  In result, students could not
check themselves the potential feedbacks from the tester scripts.  With
this change, the submissions details are available but the contents is
filtered, so neither submission can be accessed nor student comments
can be read until the deadline is reached.