Add an extra check at get-submission for ballot box submissions.
Now that it is possible to access the submission details even in ballot
box mode, the user may aquire enough information to infer the
get-submission link and download the body of the submission (even if
that is missing from the page).  In order to avoid such tricky cases, add
an extra check to the corresponding user story action so it is explicitly
verified if it is actually allowed for the user to download the requested
file or not.